requirements.txt
dnspython
python-whois
jsonschema
datetime